add-cliendIssueReport/Deploy-v1

+++ app/api/dass21/[id]/route.ts
@@ -1,14 +1,26 @@
+/* eslint-disable @typescript-eslint/no-explicit-any, @typescript-eslint/no-wrapper-object-types, @typescript-eslint/no-unused-vars */
+type SegmentParams<T extends Object = any> = T extends Record<string, any>
+    ? { [K in keyof T]: T[K] extends string ? string | string[] | undefined : never }
+    : T
+
+type RouteContext = { params: SegmentParams }
+
+
 import { NextResponse } from "next/server";
 import prisma from "@/utils/db";
 
-// Add `request: Request` as the first argument
-export async function DELETE(request: Request, { params }: { params: { id: string } }) {
+export async function DELETE(
+    req: Request,
+    context: RouteContext
+) {
     try {
-        const id = parseInt(params.id);
+        const id = Number(context.params.id);
+        if (isNaN(id)) {
+            return NextResponse.json({ error: "Invalid ID" }, { status: 400 });
+        }
+
         await prisma.dass_21_result.delete({
-            where: {
-                id: id
-            }
+            where: { id }
         });
 
         return NextResponse.json({ message: "Delete User Success !" }, { status: 200 });
@@ -16,8 +28,8 @@ export async function DELETE(request: Request, { params }: { params: { id: strin
         if (error instanceof Error) {
             console.error("DELETE ID DASS21 Error : ", error.message);
         } else {
-            console.error("Unknow error in DELETE ID DASS21 ! ", error);
+            console.error("Unknown error in DELETE ID DASS21 ! ", error);
         }
-        return NextResponse.json({ message: "Server DELETE ID DASS21 Error!" }, { status: 500 }); // Use 500 for server errors
+        return NextResponse.json({ message: "Server DELETE ID DASS21 Error!" }, { status: 500 });
     }
 }
